--- a/db/migrate/20250128101515_add_edition_link_index.rb
+++ b/db/migrate/20250128101515_add_edition_link_index.rb
@@ -1,0 +1,5 @@
+class AddEditionLinkIndex < ActiveRecord::Migration[8.0]
+  def change
+    add_index :links, %i[edition_id link_type]
+  end
+end