--- a/spec/graphql/sources/linked_to_editions_source_spec.rb
+++ b/spec/graphql/sources/linked_to_editions_source_spec.rb
@@ -33,4 +33,25 @@ RSpec.describe Sources::LinkedToEditionsSource do
       expect(request.load).to eq([target_edition_1, target_edition_3])
     end
   end
+
+  it "returns a mixture of links when both present" do
+    target_edition_1 = create(:edition)
+    target_edition_2 = create(:edition)
+    target_edition_3 = create(:edition)
+
+    source_edition = create(:edition,
+                            links_hash: {
+                              "test_link" => [target_edition_1.content_id],
+                              "another_link_type" => [target_edition_2.content_id],
+                            })
+
+    link_set = create(:link_set, content_id: source_edition.content_id)
+    create(:link, link_set: link_set, target_content_id: target_edition_3.content_id, link_type: "test_link")
+
+    GraphQL::Dataloader.with_dataloading do |dataloader|
+      request = dataloader.with(described_class, parent_object: source_edition).request("test_link")
+
+      expect(request.load).to eq([target_edition_1, target_edition_3])
+    end
+  end
 end